requirements.txt
fastapi==0.115.8
grpcio==1.70.0
grpcio-tools==1.70.0
tenacity==9.0.0
requests==2.32.3
uvicorn==0.34.0
hazelcast-python-client==5.3.0